fixed warning

+++ modules/3d/src/rgbd/odometry_frame_impl.cpp
@@ -219,7 +219,7 @@ template<typename TMat>
 void OdometryFrameImplTMat<TMat>::setPyramidAt(InputArray  _img, OdometryFramePyramidType pyrType, size_t level)
 {
     CV_Assert(pyrType >= 0);
-    CV_Assert(pyrType < pyramids.size());
+    CV_Assert(pyrType < (int)pyramids.size());
     CV_Assert(level < pyramids[pyrType].size());
     TMat img = getTMat<TMat>(_img);
     pyramids[pyrType][level] = img;
